requirements.txt
pytesseract
Pillow
pynput
googletrans==3.1.0a0
pyautogui